+++ docs/release-notes.rst
@@ -33,6 +33,11 @@ This document provides the release notes of the ricsdl library.
 Version history
 ---------------
 
+[2.0.1] - 2020-01-08
+
+* Add a fake database backend implementation to be used only for testing
+  purposes when the real DBAAS database service is not available.
+
 [2.0.0] - 2020-01-03
 
 * Change find_keys() and find_and_get() API functions to support glob-style